+++ routes/Mass.js
@@ -22,7 +22,7 @@ class Mass {
     listOfCommands() {
         var that = this;
 
-        this._program.command('masspro template', '[filename]')
+        this._program.command('masspro template user', '[filename]')
         .description("Download the csv template for importing users")
         .option('-v, --verbose', 'Use verbose console mode')
         .on('-h, --help', function(){
@@ -39,7 +39,33 @@ class Mass {
         .action(function(filename, commands) {
 
             var options = {
-                csv: filename || 'template.csv'
+                "csv": filename || 'template.csv',
+                "type": "user"
+            };
+
+            Logger.isActive = commands.verbose || false;
+            that._masspro.template(options); 
+        });
+
+        this._program.command('masspro template device', '[filename]')
+        .description("Download the csv template for importing devices")
+        .option('-v, --verbose', 'Use verbose console mode')
+        .on('-h, --help', function(){
+            console.log('  Examples:');
+            console.log('');
+            console.log("    $ rbw masspro template");
+            console.log("    $ rbw masspro template template.csv");
+            console.log('');
+            console.log('');
+            console.log('  Details:');
+            console.log('');
+            console.log('    If no filename is provided, a default template.csv file is created in the current directory.');
+        })
+        .action(function(filename, commands) {
+
+            var options = {
+                "csv": filename || 'template.csv',
+                "type": "device"
             };
 
             Logger.isActive = commands.verbose || false;